refactor: combine open and save dialogs
