+++ src/command/mod.rs
@@ -23,7 +23,7 @@ pub fn batch<M>(commands: impl IntoIterator<Item = Command<M>>) -> Command<M> {
     Command::batch(commands)
 }
 
-/// Yields a command which will run the future on thet runtime executor.
+/// Yields a command which will run the future on the runtime executor.
 pub fn future<M: Send + 'static>(future: impl Future<Output = M> + Send + 'static) -> Command<M> {
     Command::single(Action::Future(Box::pin(future)))
 }
